+++ SCrawler/API/Instagram/SiteSettings.vb
@@ -446,10 +446,11 @@ Namespace API.Instagram
         Friend Overrides Sub EndInit()
             Try : MyLastRequests.Add(LastDownloadDate.Value, LastRequestsCount.Value) : Catch : End Try
             If Not CBool(HH_IG_WWW_CLAIM_USE.Value) Then Responser.Headers.Remove(Header_IG_WWW_CLAIM)
-            If CInt(SettingsVersion.Value) < 1 Then
-                SettingsVersion.Value = 1
-                HH_IG_WWW_CLAIM_UPDATE_INTERVAL.Value = 59
-                HH_IG_WWW_CLAIM_USE_DEFAULT_ALGO.Value = 0
+            If CInt(SettingsVersion.Value) < 2 Then
+                SettingsVersion.Value = 2
+                HH_IG_WWW_CLAIM_UPDATE_INTERVAL.Value = 120
+                HH_IG_WWW_CLAIM_USE_DEFAULT_ALGO.Value = 1
+                HH_IG_WWW_CLAIM_RESET_EACH_TARGET.Value = 0
             End If
             MyBase.EndInit()
         End Sub